fix error

+++ dao/base.go
@@ -96,7 +96,7 @@ func (bd *BaseDao) buildQueryObject(dbClient *godb.DBClient, table string, optio
 		whereSQL = whereSQL + " )"
 	}
 
-	return dbClient.GormDB.Table(table).Where(whereSQL, bindDataList...).Limit(option.Size).Offset((option.Page - 1) * option.Size), nil
+	return dbClient.GormDB.Table(table).Where(whereSQL, bindDataList...).Limit(option.Size).Offset(int64(option.Page-1) * option.Size), nil
 }
 
 // getBatchCreateSQL 获取批量写入数据的sql
